feat(notifications): render real WhatsApp template message across inbox + student timeline
Extract the WhatsApp template-rendering (used by the WhatsApp Inbox) into a shared
WhatsAppTemplateRenderer and reuse it in the per-student communication timeline, so all
read paths rebuild the actual message from one source of truth. The student Notifications
tab now shows the substituted text instead of the opaque 'WhatsApp Template: ... | Params'
summary, and marks failed sends as FAILED instead of a blanket DELIVERED. Rendering is
additive: non-template / legacy payload shapes keep the existing parse.
